pyproject.toml
[project]
name = "cahn-hilliard-spectral-solver"
version = "0.1.0"
description = "Spectral IMEX Cahn-Hilliard solver with ternary Flory-Huggins thermodynamics"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.0",
    "scipy>=1.13",
    "matplotlib>=3.9",
    "hydra-core>=1.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0",
    "ruff>=0.9",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
